Update run-example
Delaying while OpenAMASE opens appears to be important. It takes a while for OpenAMASE to actually open, and the user is supposed to click on the run button before OpenUxAS instances start. The five seconds offered by some examples is insufficient, and may not be guaranteed in any case.

We make the delay configurable, so that the user can remove or adjust it, as needed.

We also make sure that we kill off forked subprocesses and report if we're unable to do so.