Fix run-tests when coverage is not available
We were using the presence of the UXAS_SOURCE_DIR and UXAS_BUILD_DIR
variables to tell us that coverage information had been computed during
the uxas build - but that is not precise. Those variables are now always
set by `anod printenv`. The fix is to check for the presence of files
generated by gcov, which is more precise.